+++ src/main/java/com/telepathicgrunt/repurposedstructures/misc/MobMapTrades.java
@@ -1,6 +1,8 @@
 package com.telepathicgrunt.repurposedstructures.misc;
 
+import com.telepathicgrunt.repurposedstructures.RepurposedStructures;
 import com.telepathicgrunt.repurposedstructures.modinit.RSStructures;
+import net.fabricmc.fabric.api.object.builder.v1.trade.TradeOfferHelper;
 import net.minecraft.item.map.MapIcon;
 import net.minecraft.village.TradeOffers;
 import net.minecraft.village.VillagerProfession;
@@ -11,32 +13,32 @@ public class MobMapTrades {
 
     public static void addMapTrades(){
 
-        // DO NOT USE Fabric API's TradeOfferHelper.registerVillagerOffers or TradeOfferHelper.registerWanderingTraderOffers
-        // It is broken and will duplicate older entries like mad. And it breaks any mod that doesn't use the API to add trades.
-
         // Villagers
         // Level 3 trades
-        addTrade(VillagerProfession.CARTOGRAPHER, 3, new TradeOffers.SellMapFactory(13, RSStructures.FORTRESS_JUNGLE, MapIcon.Type.BANNER_GREEN, 12, 5));
+        if(RepurposedStructures.RSAllConfig.RSFortressesConfig.jungleFortress.jungleFortressAverageChunkDistance != 1001)
+            TradeOfferHelper.registerVillagerOffers(VillagerProfession.CARTOGRAPHER, 3, (factories) -> factories.add(new TradeOffers.SellMapFactory(13, RSStructures.FORTRESS_JUNGLE, MapIcon.Type.BANNER_GREEN, 12, 5)));
         
         // Level 4 trades
-        addTrade(VillagerProfession.CARTOGRAPHER, 4, new TradeOffers.SellMapFactory(14, RSStructures.FORTRESS_JUNGLE, MapIcon.Type.BANNER_GREEN, 12, 10));
-        addTrade(VillagerProfession.CARTOGRAPHER, 4, new TradeOffers.SellMapFactory(14, RSStructures.MANSION_BIRCH, MapIcon.Type.MANSION, 12, 10));
-        addTrade(VillagerProfession.CARTOGRAPHER, 4, new TradeOffers.SellMapFactory(14, RSStructures.MANSION_DESERT, MapIcon.Type.MANSION, 12, 10));
-        addTrade(VillagerProfession.CARTOGRAPHER, 4, new TradeOffers.SellMapFactory(14, RSStructures.MANSION_JUNGLE, MapIcon.Type.MANSION, 12, 10));
-        addTrade(VillagerProfession.CARTOGRAPHER, 4, new TradeOffers.SellMapFactory(14, RSStructures.MANSION_OAK, MapIcon.Type.MANSION, 12, 10));
-        addTrade(VillagerProfession.CARTOGRAPHER, 4, new TradeOffers.SellMapFactory(14, RSStructures.MANSION_SAVANNA, MapIcon.Type.MANSION, 12, 10));
-        addTrade(VillagerProfession.CARTOGRAPHER, 4, new TradeOffers.SellMapFactory(14, RSStructures.MANSION_SNOWY, MapIcon.Type.MANSION, 12, 10));
-        addTrade(VillagerProfession.CARTOGRAPHER, 4, new TradeOffers.SellMapFactory(14, RSStructures.MANSION_TAIGA, MapIcon.Type.MANSION, 12, 10));
-    }
+        if(RepurposedStructures.RSAllConfig.RSFortressesConfig.jungleFortress.jungleFortressAverageChunkDistance != 1001)
+            TradeOfferHelper.registerVillagerOffers(VillagerProfession.CARTOGRAPHER, 4, (factories) -> factories.add(new TradeOffers.SellMapFactory(14, RSStructures.FORTRESS_JUNGLE, MapIcon.Type.BANNER_GREEN, 12, 10)));
 
-    public static void addTrade(VillagerProfession profession, int level, TradeOffers.Factory trade) {
-        TradeOffers.Factory[] fixedTrades = PROFESSION_TO_LEVELED_TRADE.get(profession).get(level);
-        int newSize = fixedTrades.length + 1;
+        if(RepurposedStructures.RSAllConfig.RSMansionsConfig.mansionBirchAverageChunkDistance != 1001)
+            TradeOfferHelper.registerVillagerOffers(VillagerProfession.CARTOGRAPHER, 4, (factories) -> factories.add(new TradeOffers.SellMapFactory(14, RSStructures.MANSION_BIRCH, MapIcon.Type.MANSION, 12, 10)));
+        if(RepurposedStructures.RSAllConfig.RSMansionsConfig.mansionDesertAverageChunkDistance != 1001)
+            TradeOfferHelper.registerVillagerOffers(VillagerProfession.CARTOGRAPHER, 4, (factories) -> factories.add(new TradeOffers.SellMapFactory(14, RSStructures.MANSION_DESERT, MapIcon.Type.MANSION, 12, 10)));
+        if(RepurposedStructures.RSAllConfig.RSMansionsConfig.mansionJungleAverageChunkDistance != 1001)
+            TradeOfferHelper.registerVillagerOffers(VillagerProfession.CARTOGRAPHER, 4, (factories) -> factories.add(new TradeOffers.SellMapFactory(14, RSStructures.MANSION_JUNGLE, MapIcon.Type.MANSION, 12, 10)));
+        if(RepurposedStructures.RSAllConfig.RSMansionsConfig.mansionOakAverageChunkDistance != 1001)
+            TradeOfferHelper.registerVillagerOffers(VillagerProfession.CARTOGRAPHER, 4, (factories) -> factories.add(new TradeOffers.SellMapFactory(14, RSStructures.MANSION_OAK, MapIcon.Type.MANSION, 12, 10)));
+        if(RepurposedStructures.RSAllConfig.RSMansionsConfig.mansionSavannaAverageChunkDistance != 1001)
+            TradeOfferHelper.registerVillagerOffers(VillagerProfession.CARTOGRAPHER, 4, (factories) -> factories.add(new TradeOffers.SellMapFactory(14, RSStructures.MANSION_SAVANNA, MapIcon.Type.MANSION, 12, 10)));
+        if(RepurposedStructures.RSAllConfig.RSMansionsConfig.mansionSnowyAverageChunkDistance != 1001)
+            TradeOfferHelper.registerVillagerOffers(VillagerProfession.CARTOGRAPHER, 4, (factories) -> factories.add(new TradeOffers.SellMapFactory(14, RSStructures.MANSION_SNOWY, MapIcon.Type.MANSION, 12, 10)));
+        if(RepurposedStructures.RSAllConfig.RSMansionsConfig.mansionTaigaAverageChunkDistance != 1001)
+            TradeOfferHelper.registerVillagerOffers(VillagerProfession.CARTOGRAPHER, 4, (factories) -> factories.add(new TradeOffers.SellMapFactory(14, RSStructures.MANSION_TAIGA, MapIcon.Type.MANSION, 12, 10)));
 
-        TradeOffers.Factory[] newTrades = new TradeOffers.Factory[newSize];
-        System.arraycopy(fixedTrades, 0, newTrades, 0, fixedTrades.length);
-        newTrades[newSize - 1] = trade;
-
-        PROFESSION_TO_LEVELED_TRADE.get(profession).put(level, newTrades);
+        // Wandering Traders
+        if(RepurposedStructures.RSAllConfig.RSBastionsConfig.bastionUndergroundAverageChunkDistance != 10001)
+            TradeOfferHelper.registerWanderingTraderOffers(2, (factories) -> factories.add(new TradeOffers.SellMapFactory(42, RSStructures.BASTION_UNDERGROUND, MapIcon.Type.MANSION, 1, 100)));
     }
 }